Funkční role, přidání úprava uživatelů

+++ backend/Backend/Backend/Controllers/CommentsController.cs
@@ -6,6 +6,7 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Backend.Models;
+using Backend.Utils;
 
 namespace Backend.Controllers
 {
@@ -24,35 +25,40 @@ namespace Backend.Controllers
         [HttpGet]
         public async Task<ActionResult<IEnumerable<CommentDTO>>> GetComment()
         {
-          if (_context.Comment == null)
-          {
-              return NotFound();
-          }
+            if (_context.Comment == null)
+            {
+                return NotFound();
+            }
             return await _context.Comment
                 .OrderByDescending(x => x.Created)
                 .Include(x => x.User)
                 .Select(x => new CommentDTO
-                    {
-                        UserId = x.UserId,
-                        VideoId = x.VideoId,
-                        Text = x.Text,
-                        Created = x.Created,
-                        User = x.User.ToDTO()
-                    })
+                {
+                    UserId = x.UserId,
+                    VideoId = x.VideoId,
+                    Text = x.Text,
+                    Created = x.Created,
+                    User = x.User.ToDTO()
+                })
                 .ToListAsync();
         }
 
         // PUT: api/Comments/5
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPut("{id}")]
-        public async Task<IActionResult> PutComment(Guid id, Comment comment)
+        public async Task<IActionResult> PutComment(Guid id, CommentPutDTO commentDTO)
         {
-            if (id != comment.UserId)
+            if (_context.Comment == null)
             {
-                return BadRequest();
+                return NotFound();
             }
-
+            var comment = await _context.Comment.FindAsync(id);
+            if (comment == null)
+            {
+                return NotFound();
+            }
             _context.Entry(comment).State = EntityState.Modified;
+            comment.Text = commentDTO.Text;
 
             try
             {
@@ -60,14 +66,7 @@ namespace Backend.Controllers
             }
             catch (DbUpdateConcurrencyException)
             {
-                if (!CommentExists(id))
-                {
-                    return NotFound();
-                }
-                else
-                {
-                    throw;
-                }
+                throw;
             }
 
             return NoContent();
@@ -76,12 +75,25 @@ namespace Backend.Controllers
         // POST: api/Comments
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPost]
-        public async Task<ActionResult<Comment>> PostComment(Comment comment)
+        public async Task<ActionResult> PostComment(CommentPostDTO commentDTO)
         {
-          if (_context.Comment == null)
-          {
-              return Problem("Entity set 'MyDbContext.Comment'  is null.");
-          }
+            if (_context.Comment == null)
+            {
+                return Problem("Entity set 'MyDbContext.Comment'  is null.");
+            }
+            var userId = User.GetUserId();
+            if (userId == null)
+            {
+                return Unauthorized();
+            }
+
+            var comment = new Comment()
+            {
+                Created = DateTime.Now,
+                Text = commentDTO.Text,
+                VideoId = commentDTO.VideoId,
+                UserId = (Guid)userId
+            };
             _context.Comment.Add(comment);
             try
             {
@@ -89,17 +101,10 @@ namespace Backend.Controllers
             }
             catch (DbUpdateException)
             {
-                if (CommentExists(comment.UserId))
-                {
-                    return Conflict();
-                }
-                else
-                {
-                    throw;
-                }
+                throw;
             }
 
-            return CreatedAtAction("GetComment", new { id = comment.UserId }, comment);
+            return Ok();
         }
 
         // DELETE: api/Comments/5
@@ -115,16 +120,17 @@ namespace Backend.Controllers
             {
                 return NotFound();
             }
+            var userId = User.GetUserId();
+            if (userId == null)
+            {
+                return Unauthorized();
+            }
+            //if (comment.UserId )
 
             _context.Comment.Remove(comment);
             await _context.SaveChangesAsync();
 
             return NoContent();
-        }
-
-        private bool CommentExists(Guid id)
-        {
-            return (_context.Comment?.Any(e => e.UserId == id)).GetValueOrDefault();
         }
     }
 }
